ssz: rename ssz sos serializers to shorter names
