+++ ssz/src/main/java/org/ethereum/beacon/ssz/visitor/SosDeserializer.java
@@ -7,7 +7,7 @@ import org.ethereum.beacon.ssz.type.SSZCompositeType;
 import org.ethereum.beacon.ssz.type.SSZContainerType;
 import org.ethereum.beacon.ssz.type.SSZListType;
 import org.ethereum.beacon.ssz.type.SSZType;
-import org.ethereum.beacon.ssz.visitor.SSZSosDeserializer.DecodeResult;
+import org.ethereum.beacon.ssz.visitor.SosDeserializer.DecodeResult;
 import org.javatuples.Pair;
 
 import java.nio.ByteOrder;
@@ -18,7 +18,7 @@ import java.util.function.Consumer;
 /**
  * SSZ deserializer with offset-based decoding of variable sized elements
  */
-public class SSZSosDeserializer implements SSZVisitor<DecodeResult, Pair<Bytes, Boolean>> {
+public class SosDeserializer implements SSZVisitor<DecodeResult, Pair<Bytes, Boolean>> {
   static final int BYTES_PER_LENGTH_OFFSET = 4;
 
   /**
